requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
openai>=1.12.0
anthropic>=0.18.0
google-genai>=0.1.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
